+++ tdoc/common/ext/math.py
@@ -1,17 +1,9 @@
 # Copyright 2025 Remy Blank <remy@c-space.org>
 # SPDX-License-Identifier: MIT
 
-import re
-
-from docutils import nodes
-from docutils.parsers.rst import directives
-import pyjson5
 from sphinx.util import logging
 
 from . import __version__, Dyn, dyn, tdoc_config
-from .. import util
-
-# TODO: Move :template: arguments to directive content
 
 _log = logging.getLogger(__name__)
 
@@ -27,39 +19,17 @@ def setup(app):
     }
 
 
-template_re = re.compile(r'(?s)([a-zA-Z0-9_-]+)(?:\((.*)\))?')
+class JsxGraph(Dyn):
+    has_content = True
+    has_templates = True
 
-
-class TemplatedDyn(Dyn):
-    optional_arguments = 1
-    option_spec = Dyn.option_spec | {
-        'template': directives.unchanged,
-    }
-
-    def populate(self, node):
-        if (v := self.options.get('template')) is not None:
-            if 'name' in node:
-                raise Exception(f"{{{self.name}}} Directive with :template: "
-                                "must not have a name")
-            if (m := template_re.fullmatch(v.strip())) is None:
-                raise Exception(
-                    f"{{{self.name}}} Invalid :template: value: {v}")
-            attrs = {'data-template': m.group(1)}
-            if (v := m.group(2)) is not None:
-                attrs['data-args'] = util.to_json(pyjson5.decode(f'[{v}]'))
-            node['attrs'] = attrs
-        elif 'name' not in node:
-            raise Exception(f"{{{self.name}}} Directive must have a name")
-        super().populate(node)
-
-
-class JsxGraph(TemplatedDyn):
     def populate(self, node):
         node['classes'].append('jxgbox')
-        super().populate(node)
 
 
-class ChartJs(TemplatedDyn): pass
+class ChartJs(Dyn):
+    has_content = True
+    has_templates = True
 
 
 def add_css_js(app, page, template, context, doctree):