+++ docs/release-notes.md
@@ -113,7 +113,7 @@
 
 - Added the {rst:role}`hspace` and {rst:role}`vspace` roles.
 - Added the {rst:role}`leader` role.
-- Added the {rst:dir}`{jsxgraph} :template: <jsxgraph:template>` option.
+- Added the `{jsxgraph} :template:` option.
 - [Full changelog](https://github.com/t-doc-org/common/compare/0.60...0.61)
 
 (release-0-60)=